requirements.txt
PySide6-Fluent-Widgets
pymongo[srv]
bcrypt
boto3
